Let me start work on the char RNN.
It is hard, but I have to continue for Singularity's sake. It is not going to happen on its own. The Danganronpa 3 playthrough can wait.

Let me just do this thing for today. I will take the old test and dust it off and make the MI-RNN.

The process of making that step is the same as usual. Gather the desire and manifest it into code.

One day, that desire will be great enough to draw in whole universe and the code powerful enough to be carved into reality. Without that hope, there would be no point in programming. It would just be another job. If it fell down to that then there would be no reason to do it over anything else.

In today's world, programming is certainly not a high status occupation.

That will change some day.

That will change when someone comes along and does it.