``` inl truncate network s =     inl s = s.RegionMem.create     inl network =         Struct.map (function             | {state} as d -> {d without bck with state = Struct.map (inl x -> x.update_body (inl {x with ar} -> s.RegionMem.assign ar.ptr; x)) (primals state) |> heap}             | d -> {d without bck}             ) network     {network s}
inl train {data={input label} network learning_rate final} s = // TODO: Work in progress.
    inl s = s.RegionMem.create
    inl range = fst input.dim
    assert (range = fst label.dim) "The input and label must have the same outer(1) dimension."
    Loops.for' {range with state={cost=dyn 0.0; network s};
        body=inl {i next state} ->
            s.refresh

            inl input, label = input i, label i
            inl range = fst input.dim
            assert (range = fst label.dim) "The input and label must have the same outer(2) dimension."

            Loops.for' {range with state={cost network s};
                body=inl {i next state={d with network s}} ->
                    inl input, label = input i, label i
                    inl network, input = run s input network
                    inl {out bck} = final label input s
                    inl prev =
                        inl bck = Struct.map (inl {bck} -> bck) network, bck
                        inl x = heap {out bck}
                        dyn match d with {prev} -> List.cons x prev | _ -> List.singleton x
                    inl network = Struct.map (inl x -> {x without bck})
                    next {d with prev network}
                finally=inl {state with cost network s prev} ->
                    List.foldr (inl {bck} _ ->
                        inl learning_rate = learning_rate ** 0.85f32
                        inl apply bck = bck {learning_rate} |> ignore
                        Struct.foldr (inl {bck} -> Struct.foldr (inl bck _ -> apply bck) bck) network ()
                        ) prev ()

                    Optimizer.standard learning_rate s network

                    inl cost = List.foldr (inl {out} cost -> cost + (s.CudaTensor.get out |> to float64)) prev cost

                    inl {network s} = truncate network s
                    if nan_is cost then s.RegionMem.clear; cost
                    else next {cost network s}
                }
        finally=inl {cost s} -> s.RegionMem.clear; cost
        }
    |> inl cost -> cost / to float64 input.span_outer2
```

Should this be fine now? I am thinking of how I could abstract this and am having difficulty. Well, that was the case last time I tried this. Functional programming can only go so far to make up for a lack of garbage collection. These two loops are going to have to be a composite and going from his to a sampler will simply have to involve copy pasting. It is futile to try and reduce this.

It would be possible as in the last library, but the real issue is that it is necessary to reason about transfer timings and memory allocation. Abstraction does not help there and in fact I'd argue that separating this function would make those two aspects even harder than they already are.

Honestly, I am not as bothered with this as before. These tests will simply have to be thrown away if I ever change the architecture of the library again though I do not imagine I will need to.

I've spent the last 3 hours essentially trying to eat the above. I'll just have to internalize it and then move to the next thing.

Let me step away from the screen to rest for a bit. After that I will deal with the test.