I really was away for a while. Async training, resnets, resnets with PRONG - those are going to be a nightmare to optimize. But I can do it with Spiral.
I've been thinking about async training and doing async updates of weights...does not really sit well with me at all. What I will do instead is use the wavefront iteration pattern. When I am doing that splitting the layer into multiple ones will require special considerations in the `run` function because the lower layers might need to wait on the upper ones for `state`.

It will be hard, but I can think of solutions for all these problems. I will be able to handle it when the time comes. It won't be as simple as stacking layers and will require quite a bit of work, but I'll make it work out.

My effort will see me through.

This char RNN test is a pain. Well, I knew that making this loop would be hard which is why I wanted to avoid it. Let me finish it and maybe I'll call it a day.