``` Starting timing for: iteration 0 The time was 00:00:05.2398641 for: iteration 0 Training: 189.140459140434 Starting timing for: iteration 1 The time was 00:00:05.2065724 for: iteration 1 Training: 166.960033630624 Starting timing for: iteration 2 The time was 00:00:05.1970832 for: iteration 2 Training: 162.910574999802 Starting timing for: iteration 3 The time was 00:00:05.2023904 for: iteration 3 Training: 161.19090795254 Starting timing for: iteration 4 The time was 00:00:05.2027856 for: iteration 4 Training: 160.228574011256 ```
Strange that it is this slow. 5.2s is how long it takes a MI layer without wavefront. And that the cost is this low, though that might be simply because I am handling the cost differently.

Well, let me suppose that I have this half done at this point. I am going to have to go to the other branch and compare a feedforward net there.