Ah, the speed is probably because I was using layers of size 128 last time. That would explain it. Let me give it a try.
```
Starting timing for: iteration 0
The time was 00:00:04.8333136 for: iteration 0
Training: 191.558674338109
Starting timing for: iteration 1
The time was 00:00:04.8412303 for: iteration 1
Training: 167.877131995033
Starting timing for: iteration 2
The time was 00:00:04.9025619 for: iteration 2
Training: 163.422805099365
Starting timing for: iteration 3
The time was 00:00:04.8572400 for: iteration 3
Training: 161.499282920624
Starting timing for: iteration 4
The time was 00:00:04.8315414 for: iteration 4
Training: 160.466795710956
```

Or maybe not. I have no idea.

But the branch I am using is not far behind. I need to go back to all the way before I started work on the new library.

Let me give it a shot.