"``` Starting timing for: iteration 0 The time was 00:00:04.3163977 for: iteration 0 Training: 198.456978405223 Starting timing for: iteration 1 The time was 00:00:04.2849198 for: iteration 1 Training: 172.196720235488 Starting timing for: iteration 2 The time was 00:00:04.2825037 for: iteration 2 Training: 166.423215305104 Starting timing for: iteration 3 The time was 00:00:04.2931790 for: iteration 3 Training: 163.639467744266 Starting timing for: iteration 4 The time was 00:00:04.3467906 for: iteration 4 Training: 162.06596363292 ```
I expected a linear layer would be a lot faster, but I guess I was completely wrong. It is at most 20% faster than the `mi` layer. And the cost is in fact supposed to be this low. Maybe I should use a feedforward network to do this instead of LSTM.

```
Starting timing for: iteration 0
The time was 00:00:05.3298445 for: iteration 0
Training: 178.979077563566
Starting timing for: iteration 1
The time was 00:00:05.2674551 for: iteration 1
Training: 144.215340221629
Starting timing for: iteration 2
The time was 00:00:05.2551754 for: iteration 2
Training: 133.739597124212
Starting timing for: iteration 3
The time was 00:00:05.2115177 for: iteration 3
Training: 126.997966373668
Starting timing for: iteration 4
The time was 00:00:05.2570856 for: iteration 4
Training: 122.29655083488
```

Here is how an RNN does. Actually, should the time be lower than this? I remember it being 4.3 per iteration.

```
Starting timing for: iteration 0
The time was 00:00:05.0289851 for: iteration 0
Training: 165.896009108599
Starting timing for: iteration 1
The time was 00:00:05.0374324 for: iteration 1
Training: 141.125713937423
Starting timing for: iteration 2
The time was 00:00:05.0269430 for: iteration 2
Training: 124.62828689463
Starting timing for: iteration 3
The time was 00:00:04.9860218 for: iteration 3
Training: 119.120690205518
Starting timing for: iteration 4
The time was 00:00:05.0043856 for: iteration 4
Training: 115.145506241742
```

Here is the MI-LN RNN.

Meh, maybe the GPU is feeling slow today. It tends to perform differently from time to time for unknown reasons. If I'd reset it would do better. At any rate, this is not worth concerning myself with. 10% differential in performance is perfectly explainable by the wavefront iteration done in the old library.

With this the test really is complete. I've confirmed that there is a match."