``` inl mutable_function_mult f {state=(!heap state) input} =     inl f = Struct.map (met f x -> f x |> inl {state out} -> {state=heap state; out=heap out}) f     inl unroll_state f state =         inl rec loop f state =             inl state' = f {state input} .state             if eq_type state state' then state             else state \/ loop f state'
        Struct.foldl (inl state f -> match state with () | _ -> loop f state) state f

    inl ty =
        type
            inl rec loop prev =
                inl cur = unroll_state f prev
                if eq_type prev cur then cur else loop cur
            loop state
```

I've been staring at this for a while now.

What is my genius plan now that I have this? Sure, this will infer all the types, but I am unsure of to deal with the RNN in the context of a mutable function. Here is what I am considering doing.

```
inl infer f {state=(!heap state) input} =
    inl f = Struct.map (met f x -> f x |> inl {state out} -> {state=heap state; out=heap out}) f
    inl unroll_state f state =
        inl rec loop f state =
            inl state' = f {state input} .state
            if eq_type state state' then state
            else state \/ loop f state'

        Struct.foldl (inl state f -> match state with () | _ -> loop f state) state f

    inl ty =
        type
            inl rec loop prev =
                inl cur = unroll_state f prev
                if eq_type prev cur then cur else loop cur
            loop state

    ty, f
```

Truncating the states will require its own thing. I am going to have to make a specialized loop for RNNs.

Let me try doing it. Getting that loop operational and completing the char RNN will be my focus for the day. It probably won't be the whole day - I'll be done with this in the next few hours. After that I guess I'll plugging PRONG into DP after trying it out on the char RNN first. But that will be easy.

This part I am working on now is the one that requires real inovation.