"...Hmmm, nope. Let me just print out cost. Why would it be zero everywhere? I am probably missing an argument somewhere but where?
```
                    inl cost =
                        List.foldl (inl cost -> function
                            | {out} -> cost + out()
                            | _ -> cost
                            ) cost prev_states

                    Console.writeline cost
```

It is zero here?

```
| {out} -> cost + out()
```

Ah whops. I get it.

```
| {prev={out}} -> cost + out()
```

It should be like this.

```
Starting timing for: iteration 0
The time was 00:00:04.8611785 for: iteration 0
Training: 191.558673652656
Starting timing for: iteration 1
The time was 00:00:04.7695128 for: iteration 1
Training: 167.877400632729
Starting timing for: iteration 2
The time was 00:00:04.8729935 for: iteration 2
Training: 163.423594316577
Starting timing for: iteration 3
The time was 00:00:04.9490318 for: iteration 3
Training: 161.499166132773
Starting timing for: iteration 4
The time was 00:00:04.8263864 for: iteration 4
Training: 160.468243919751
```

Yeah here it is.

But wow calculating the costs sure slows it down.

```
Starting timing for: iteration 0
The time was 00:00:03.4722242 for: iteration 0
Training: 0
Starting timing for: iteration 1
The time was 00:00:03.4656998 for: iteration 1
Training: 0
Starting timing for: iteration 2
The time was 00:00:03.5348069 for: iteration 2
Training: 0
Starting timing for: iteration 3
The time was 00:00:03.5951912 for: iteration 3
Training: 0
Starting timing for: iteration 4
The time was 00:00:03.6202645 for: iteration 4
Training: 0
```

I just commented out the part of the code that sums up the costs. Those device to host transfers slow things down by a significant amount. 33%.

Well, nevermind that. Let me finally test out the RNN. It has all been leading up to this.

Finally I get to do it."