``` Starting timing for: iteration 0 The time was 00:00:06.3278977 for: iteration 0 Training: 162.967074545429 Starting timing for: iteration 1 The time was 00:00:06.3039768 for: iteration 1 Training: 118.558704886805 Starting timing for: iteration 2 The time was 00:00:06.4063687 for: iteration 2 Training: 109.59115435183 Starting timing for: iteration 3 The time was 00:00:06.2961849 for: iteration 3 Training: 105.352582325611 Starting timing for: iteration 4 The time was 00:00:06.1106358 for: iteration 4 Training: 102.942497201045 ```
It took some tuning of learning rates, but this is way better than what I had in the other library. The reason why I was not getting good results there is because I used a wider grid search. Well, forget that.

it does not matter anymore. This is really quite good.

I wonder how better it will get once I replaced the regular matrix multiplies with natural ones. Also this is what I am going to rename `naturalize` to - `natural_matmultb`.

Let me give it a try.