Oho. Increasing the learning rate to `2 ** -12` is doing wonders for it.
```
Starting timing for: iteration 0
The time was 00:00:25.5258279 for: iteration 0
Training: 136.901730146478
Starting timing for: iteration 1
The time was 00:00:25.5683212 for: iteration 1
Training: 107.918182862594
Starting timing for: iteration 2
The time was 00:00:25.8627844 for: iteration 2
Training: 104.877469058861
Starting timing for: iteration 3
The time was 00:00:26.1665389 for: iteration 3
Training: 103.350226365906
Starting timing for: iteration 4
The time was 00:00:26.0061949 for: iteration 4
Training: 102.643002372673
```

Since even PRONG is having trouble going lower than this maybe this is the limit of the network.

Let me increase the learning rate by 2x once more.

Also, I forgot that I am not using prong in the decoder. I'll try doing that later.