``` Starting timing for: iteration 0 The time was 00:00:34.5227570 for: iteration 0 Training: 125.342853918672 Starting timing for: iteration 1 The time was 00:00:34.8916465 for: iteration 1 Training: 109.468823843581 Starting timing for: iteration 2 The time was 00:00:34.4614555 for: iteration 2 Training: 108.076148468344 Starting timing for: iteration 3 The time was 00:00:34.4979079 for: iteration 3 Training: 107.531392182717 Starting timing for: iteration 4 The time was 00:00:34.3537982 for: iteration 4 Training: 106.938438221593 ```
This is with `2 ** -11.5`. Using PRONG in the decoder really spikes the time. It would be faster to actually use SGD for SL. But for RL it does not matter. PRONG is worth it.

Right now it seems I am getting that effect where lowering the learning rate makes it run better. I'll try going to -12 and -12.5 next.