``` Starting timing for: iteration 0 The time was 00:00:10.9045876 for: iteration 0 Training: 172.563579952016 Starting timing for: iteration 1 The time was 00:00:11.6735745 for: iteration 1 Training: 128.992319050957 Starting timing for: iteration 2 The time was 00:00:10.7033599 for: iteration 2 Training: 117.773870327893 Starting timing for: iteration 3 ```
The old library was not that shitty. But it true that the architectures are different.

Let me implement the old architecture then in that case.

Edit:

```
                {
                b1 = Initializer.constant {dim=size; init=1f32}
                b2 = Initializer.constant {dim=size; init=0.5f32}
                b3 = Initializer.constant {dim=size; init=0.5f32}
                b4 = Initializer.constant {dim=size; init=0f32}
                input = Initializer.tanh (sublayer_size, size)
                state = Initializer.tanh (size, size)
                }
```

Ahhh...I get it. If the biases are always zero then with multiple layers no inputs will get through. Whops.

Forget implementing that other architecture. What I need to is set the biases for state at least to 1.