"8:10pm.
```
Starting timing for: iteration 0
The time was 00:01:07.2669303 for: iteration 0
Training: 158.27523265034
Starting timing for: iteration 1
The time was 00:01:08.3474424 for: iteration 1
Training: 152.428868930568
Starting timing for: iteration 2
```

Weird. It is working now. Since all allocations are done through the allocator maybe this is some kind of weird startup bug?

8:10pm. Wow, it is really super tedious to wait an entire minute for it to go through a single iteration. This is not really acceptable to me.

8:15pm.

```
Starting timing for: iteration 0
The time was 00:01:09.0771290 for: iteration 0
Training: 117.323516116423
Starting timing for: iteration 1
The time was 00:01:09.5814312 for: iteration 1
Training: 100.44240555715
Starting timing for: iteration 2
The time was 00:01:09.8469336 for: iteration 2
Training: 97.6649281842744
Starting timing for: iteration 3
The time was 00:01:10.2180846 for: iteration 3
Training: 96.5249163948876
Starting timing for: iteration 4
The time was 00:01:09.9242473 for: iteration 4
Training: 95.6609012164614
```

Damn it is slow, but one shot of this is with a learning rate of `2 ** -13` is pretty good. It seems the more layers get added the lower the learning rates need to be pulled.

I want to try replacing the linear layer with some more prong layers.

8:20pm. Right now I am trying 4 MI layers. This should have it take 2-2.5 minutes per iteration.

One thing I am really wondering about is whether I should move up the learning rate after the first iteration.

8:25pm.

```
//Loops.for' {from=0; near_to=5; body=inl {i next} -> f learning_rate}
Struct.foldr (inl learning_rate next i -> f learning_rate i next) learning_rate
```

I quick recursive fold should do the trick. Right now it is taking 1:43s to do a single iteration. Not as slow as I expected, but pretty slow.

8:30pm.

```
Starting timing for: iteration 0
The time was 00:01:42.7814873 for: iteration 0
Training: 113.905462852734
Starting timing for: iteration 1
The time was 00:01:43.3175755 for: iteration 1
Training: 98.1942006594994
Starting timing for: iteration 2
The time was 00:01:43.6356756 for: iteration 2
Training: 95.303122803788
Starting timing for: iteration 3
The time was 00:01:44.3629722 for: iteration 3
Training: 93.4837563364383
Starting timing for: iteration 4
The time was 00:01:42.9635696 for: iteration 4
Training: 92.1977560022736
```

This is all with -13.

8:35pm.

```
Error trace on line: 149, column: 5 in file "learning2".
    if nan_is cost then Console.writeline "Training diverged. Aborting..."
    ^
Error trace on line: 150, column: 10 in file "learning2".
    else next (i+1)
         ^
Invalid use of apply. [] and lit 3i64
```

I am really not fresh enough to do these recursive folds right now? What is wrong here?

8:40pm. `Struct.foldr` does not work, but `Tuple.foldr` does. This is something to investigate tomorrow. It is probably a library bug.

8:45pm. Well, since it is running I might as well do it now.

```
inl foldr f a s =
    inl rec loop s = function
        | x when caseable_box_is x -> f x s
        | x :: xs -> loop (loop s xs) x
        | () -> ()
        | {!block} & x -> module_foldr (inl k x s -> loop s x) x s
        | x -> f x s
    loop s a
```

Ah, that `| () -> ()` should be `| () -> s`.

...That mistake was in all the folds apart from `foldl`. To think there was a bug like this lying around.

Two more iterations of this just so I can see if moving up learning rates helps and I will call it a day again.

8:50pm.

```
The learning rate is 2 ** -13.00084
Starting timing for: iteration 0
The time was 00:01:44.0383818 for: iteration 0
Training: 113.889600982999
The learning rate is 2 ** -12.00083
Starting timing for: iteration 1
The time was 00:01:46.4706600 for: iteration 1
Training: 114.383445579339
The learning rate is 2 ** -12.00083
Starting timing for: iteration 2

Unhandled Exception: ManagedCuda.CudaBlas.CudaBlasException: ExecutionFailed: An access to GPU memory space failed.
```

Agh...I guess I am done. I guess two iterations is enough for today."