It is a new day and I am sitting here is my spiffing new chair. It is finally September. Only 4 months left until the next update. I really came far haven't I?
I did really well last month all things considered. With respect to RNNs I am almost completely set. Today my number one priority will definitely be to track that memory bug from yesterday. I need to figure out why it is happening. I have a couple of ideas on how to proceed.

It is too bad I did not investigate it closer yesterday, but I did look where the last error happened last time and it was in `symm`. What I will do today is run it a few more times with a larger net and see whether all of them happen in symm.

In this situation the most sensible thing would be to suspect an allocator bug, but in that case it would be strange if all the errors happened in symm.

Let me also take this chance to test depth and increasing learning rates.

This sort of testing will be time consuming as the runs have started to take a while, so I'll be doing less programming than usual. Well, the transition from me doing all the work to computer doing most of it will come eventually. This is just the first bit into it.

Once I've figured out why the bug is happening, I'll move on to testing differentiable plasticity on the char RNN. I am definitely looking forward to that. I think I'll reduce the size of the dataset by 64x since I will be doing it online.