"9:25am. I am thinking of watching some the of the PL videos by Andrej Baurej I linked here days ago. These tests are so damn time consuming. Well one more and I will be done.
```
The learning rate is 2 ** -13.00084
Starting timing for: iteration 0
The time was 00:01:40.5591397 for: iteration 0
Training: 113.889600982999
The learning rate is 2 ** -13.00084
Starting timing for: iteration 1
The time was 00:01:40.4895339 for: iteration 1
Training: 98.1818790582611
The learning rate is 2 ** -12.00083
Starting timing for: iteration 2
The time was 00:01:40.3960545 for: iteration 2
Training: 111.264207254438
```

It seems there is literally zero point in lowering learning rates. Forget super convergence.

9:30am. https://www.cs.uoregon.edu/research/summerschool/summer18/topics.php

To be honest, I am interested in all the videos here. Maybe I'll take a look at the linear lambda calculus one.

9:40am. https://www.cs.uoregon.edu/research/summerschool/summer18/lectures/downen6-1.mp4

Holy shit this is 2.6Gb.

9:45am. "From the summer school lectures, I've only looked at the first day 4 lecture by Paul Downen and it turns out the reason it is loading so slowly is because it is 2.6Gb. That is pretty insane size for a 36m video and my download budget is not unlimited. If you used better compression you could get it down to 1/5th-1/10th of that size.

Could you please do so for all the videos? Thank you."

I've actually decided to mail those guys. There is a link missing in the quote above.

One more run at -13.5 and then I am done.

9:55am. Now I am on the last iteration.

```
The learning rate is 2 ** -13.50531
Starting timing for: iteration 0
The time was 00:01:39.6024987 for: iteration 0
Training: 114.251125499168
The learning rate is 2 ** -13.50531
Starting timing for: iteration 1
The time was 00:01:39.5385730 for: iteration 1
Training: 96.8129400808145
The learning rate is 2 ** -13.50531
Starting timing for: iteration 2
The time was 00:01:38.4094138 for: iteration 2
Training: 93.8165834223961
The learning rate is 2 ** -13.50531
Starting timing for: iteration 3
The time was 00:01:38.7296009 for: iteration 3
Training: 92.1359454822891
The learning rate is 2 ** -13.50531
Starting timing for: iteration 4
The time was 00:01:39.2601047 for: iteration 4
Training: 90.9028111275505
```

Yeah, I am done with this. What I will do now is go back to a single layer and try it using online learning on 1/64th of the dataset. Then I will compare the MIRNN to the differentiable plasticity one."