``` The learning rate is 2 ** -13.50531 Starting timing for: iteration 0 The time was 00:00:06.9376059 for: iteration 0 Training: 4.77853615790167 The learning rate is 2 ** -13.50531 Starting timing for: iteration 1 The time was 00:00:06.9338704 for: iteration 1 Training: 4.59875579709735 The learning rate is 2 ** -13.50531 Starting timing for: iteration 2 The time was 00:00:06.8361622 for: iteration 2 Training: 4.39003564131118 The learning rate is 2 ** -13.50531 Starting timing for: iteration 3 The time was 00:00:06.9968719 for: iteration 3 Training: 4.15384657904232 The learning rate is 2 ** -13.50531 Starting timing for: iteration 4 The time was 00:00:07.0888132 for: iteration 4 Training: 3.93377698139589 ```
This is the Pl-RNN. I haven't done any tunning yet. To start things off, how about I find the optimal LR for MI and then for this.

Edit:

```
The learning rate is 2 ** -5.000001
Starting timing for: iteration 0
The time was 00:00:05.7925726 for: iteration 0
Training: 2.59767465548419
The learning rate is 2 ** -5.000001
Starting timing for: iteration 1
The time was 00:00:05.7946600 for: iteration 1
Training: 2.08162667306565
The learning rate is 2 ** -5.000001
Starting timing for: iteration 2
The time was 00:00:05.8685356 for: iteration 2
Training: 1.94050844411912
The learning rate is 2 ** -5.000001
Starting timing for: iteration 3
The time was 00:00:05.8969728 for: iteration 3
Training: 1.96310145758462
The learning rate is 2 ** -5.000001
Starting timing for: iteration 4
The time was 00:00:05.7816168 for: iteration 4
Training: 2.11669067689049
```

The limit for the learning rate is quite high as it turns out. This is the MIRNN. Let me the the PlRNN.