``` The learning rate is 2 ** -5.000001 Starting timing for: iteration 0 The time was 00:00:07.8243420 for: iteration 0 Training: 2.79674221518722 The learning rate is 2 ** -5.000001 Starting timing for: iteration 1 The time was 00:00:07.7524569 for: iteration 1 Training: 2.55015223719452 The learning rate is 2 ** -5.000001 Starting timing for: iteration 2 The time was 00:00:07.6434152 for: iteration 2 Training: 2.51483243339723 The learning rate is 2 ** -5.000001 Starting timing for: iteration 3 The time was 00:00:05.6410274 for: iteration 3 Training: NaN Training diverged. Aborting... ```
Let me lower the LR a tad.

```
The learning rate is 2 ** -5.500007
Starting timing for: iteration 0
The time was 00:00:07.8624695 for: iteration 0
Training: 2.82568960367515
The learning rate is 2 ** -5.500007
Starting timing for: iteration 1
The time was 00:00:07.7228882 for: iteration 1
Training: 2.54441224800586
The learning rate is 2 ** -5.500007
Starting timing for: iteration 2
The time was 00:00:07.7768792 for: iteration 2
Training: 2.4999122584715
The learning rate is 2 ** -5.500007
Starting timing for: iteration 3
The time was 00:00:07.7805149 for: iteration 3
Training: 2.48006179102274
The learning rate is 2 ** -5.500007
Starting timing for: iteration 4
The time was 00:00:07.6625793 for: iteration 4
Training: 2.46798364960555
```

This is good. Layer for layer, maybe PlRNN has lower capacity than MIRNN. Or maybe it does things simply differently. How good an architecture is cannot be judged by how well it overfits on a tiny dataset as the two RNNs have different functions. What really matters is how good their application to RL is.

What would the feedforward baseline be?