```                 fwd=inl {n x H} -> n * x + (one - n) * H                 bck=inl {in={n x H}} -> { n=x - H; x=n; H=-n } ```
Is the gradient for H really correct? No, I think it should be `one-n`.

```
                fwd=inl {n x H} -> n * x + (one - n) * H
                bck=inl {in={n x H}} -> { n=x-H; x=n; H=one-n }
```

Let me try it like this.

```
The learning rate is 2 ** -5.500007
Starting timing for: iteration 0
The time was 00:00:06.8261518 for: iteration 0
Training: 2.82565677730942
The learning rate is 2 ** -5.500007
Starting timing for: iteration 1
The time was 00:00:07.2448845 for: iteration 1
Training: 2.54439059841838
The learning rate is 2 ** -5.500007
Starting timing for: iteration 2
The time was 00:00:06.8027494 for: iteration 2
Training: 2.49988193088746
The learning rate is 2 ** -5.500007
Starting timing for: iteration 3
The time was 00:00:07.3201927 for: iteration 3
Training: 2.48001356896866
The learning rate is 2 ** -5.500007
Starting timing for: iteration 4
The time was 00:00:07.0785802 for: iteration 4
Training: 2.46775325372052
```

Ugh, lame.