This one happens rarely, but the way pattern matching is compiled can lead to surprises. Not only is Spiral overzealous in unboxing union types, but sometimes that same process can cause unintentional shadowing.
```
inl from = 1

match {from = 2} with
| {from=from'} -> print_static {from from'}
```

```
{from=lit 1i64; from'=lit 2i64}
```

This is what would be expected.

```
inl from = 1

match {from = 2} with
| {from near_to} -> () // This branch binds the from in the module and shadows the one in scope.
| {from=from'} -> print_static {from from'}
```

```
{from=lit 2i64; from'=lit 2i64}
```

Unfortunately there is no good way of fixing this. It is one of the consequences of how Spiral's patterns are compiled and there is no good alternative (as far as the author knows) to how it is done now.