I am still thinking about it. I am in the zone trying to figure out where I went wrong, but nothing comes to mind. Is one my assumptions broken?
One thing that occurs to me is that this issue might be blessing in disguise. Since it will force me to take apart the method before putting it back together, it will give me an insight into how it actually behaves. I won't be able to do it today. Most likely I will have to spend a few days experimenting with DP before I get it done.

Right now I feel like taking a look at Miconi's code to make sure that I got my assumptions correct. The first point that I need to get it is to make sure is that.

Usually in times like these it is not the case that the code I am using is broken. Rather it is the way I perceive what the model should be doing.