https://arxiv.org/pdf/1802.05098.pdf DiCE - The Infinitely Differentiable Monte Carlo Estimator
I'll read this later.

Let me gather my thoughts. I won't program anymore so let me make the closing commit now that I am done with my investigation of the differentiable plasticity code.

To start things off, let me just say that I am absolutely shocked how different the code is from the paper. Every test has a different way of passing the inputs, and the actual inputs are not passed in directly into the adaptive weight matrix like I'd expected. Instead they are added like biases. Insane. I remember once having a dream that inputs could be biases, but ruled it out as unrealistic.

Well here it is - right in the real world.

Reading the code also made me realize that I made a mistake with calculating the Hebb first so I am going to fix that. The Uber code does the right thing of doing it afterwards and passing it along.

My plan for tomorrow will be to get DP finally working. Now that I know exactly what it does that should not be a problem. I made some design decisions, but the paper really ommitted some very important details of how the architecture works. No way could have known this ahead of time.

It also opens up some questions of what the right way of passing inputs into it would be. I'll figure it out as I go along. Just adding them to the cell as biases does not sit entirely right with me. My intuition is pushing me away from that sort of thing. It will have to be something similar to multiplicative integration.

In fact combining MI and Pl could be exactly the ideal way of doing it.

Originally I thought that Pl would be a complete replacement for standard RNNs, but now I am no longer sure of that. I see it differently now. I should definitely be able to combine the two. That should allow me to draw out its true power.

I just have to keep going at it. I am so close to bringing this phase to a conclussion. At most the DP will take me two weeks to master. Then comes Rudder.

At that point I have nothing more do with respect to RNNs. It is all RL related from there. Async training of agents, NL Holdem, checkpointing and self play. This is roughly where I expected to be in mid 2016 back when it was 2015. Yes I am behind, but I started this quest in 2015 rather than 2020 for exactly this reason - to give myself time in order to set everything up.

At the end of all that come interfaces and putting them to work in the real world. That is when real magic will start.

Let me read that Dice paper I linked to at the beginning and then I'll call it a day.