https://youtu.be/xL-GKD49FXs?t=1980
"Imagine if coming to a conference and having to discuss you could just merge your minds." Well, I already expected differentiability and parameter sharing to be a great boon, so most of this talk is not that surprising to me.

Actually, I feel like resuming programming now. Let me test this on a real dataset.