Differentiable plasticity works like crap and I am bothered by all the ways of doing inputs in that I saw in the examples. A part of me wants to write to the author, but I already know how all of this needs to go.
Differentiable plasticity as it is now feels more like an isolated component than an actual unit. It feels too incomplete to be a real layer.

What I should do is take the MI layer and put differentiable plasticity into it. That will bring the two architectures on an even footing. That should immediatelly give me a clue of whether DP is a net positive or a negative. I also really should ask the author about the habit of using inputs in place of biases.

Let me fire a quick email to him and then I will call it a day. I'll put DP into MI tomorrow. If that does not work I'll just go with MI, but I want to make this work.