Actually I feel like doing just a little bit of testing.
```
    inl network =
        linear 128,
        plastic_hebb',
        linear size.hot
```

Let me do it like this. Thinking about it more if I treat the DP as a vanilla RNN, then it all makes sense. The reason why I am getting poor results here is because I am not projecting the state. Let me try this.

The way DP was presented is messing with my head for some reason. I do not understand why I am having such an overreaction to it. I guess my preconception of how it should have worked is overpowering all other considerations.