Got up a bit ago and am almost done chilling. I think I'll start it off by making a post in the PL monthly thread and then I will start programming.
Today is the day I will get this to work. And after I get this to work, I'll make optimized implementations for the weight calculations and extend it so it allow batch learning.

There is a fair bit of work there and I won't be able to do it all in one day, but it should not take me too long. A few days at most.

I have no need to rush this. Let me just do it all properly and everything will fall into place.