Yesterday I did no work for reasons unkown to me, but today it feels like I am just wasting time. I spent way too long - the usual to be honest, writing out that post on the PL sub.
Let me do it as per plan. I will extend the broadcasting_activation so it supports scalars. It is also missing bck_in_outer so I need to add that as well.

Only at a later point will I optimize the crap out of all of this.

Right now my mission is to get it to work at all.