```                 inl ty = primal ins.in_inner .elem_type                 s.CudaKernel.redo_init {                     dim                     neutral_elem=Struct.map (const zero) x_inner.elem_type; redo=Struct.map2 (+) ```
Ugh, `neutral_ELEEEEM`!!! The correct is to remove it entirely. That is the one truth.

I really hadn't counted on doing this.

I already tried doing a few times in the past, but could not reach a satisfactory conclusion.

Let me just go for it. I'll put everything else on the backburner.