What I've started just now is probably most significant change to the way GPU programming is done in Spiral in quite a while.
I've long wanted to remove having to deal with neutral elements but was unsure of how to do it until now.

I've also put in a loop elimination optimization into `grid_for_template`.

Tomorrow I will have to go through all the Cuda tests from the top, but modifying the rest of the kernels so they use the new loop should not be hard at all.

I am also not that far off from dealing with the broadcasting_activation either. I'll have to add two more cases to it, but it should not be hard now that I have the general design in mind.

As hard as it is to admit, most of the effort in redesigning the activation was in making those indexing functions and making sure everything fits together. What I had before was such a hot mess.

The way I am doing indexing now is a stroke of genius. Well, I'll call it a day here. No point in going on any further.

Hopefully I won't spend the entire morning on social media. I feel like I am getting into the swing of things now.