It definitely works better now.
Since it is not a bias addition how about I exchange that stupid thing with a hadamarad multiplication?

Edit:

```
The learning rate is 2 ** -5.500007
Starting timing for: iteration 0
The time was 00:00:09.3934858 for: iteration 0
Training: 2.66135640214794
The learning rate is 2 ** -5.500007
Starting timing for: iteration 1
The time was 00:00:09.2703308 for: iteration 1
Training: 2.13264193743613
The learning rate is 2 ** -5.500007
Starting timing for: iteration 2
The time was 00:00:09.4793108 for: iteration 2
Training: 1.93510541395656
The learning rate is 2 ** -5.500007
Starting timing for: iteration 3
The time was 00:00:09.3655695 for: iteration 3
Training: 1.82221499673356
The learning rate is 2 ** -5.500007
Starting timing for: iteration 4
The time was 00:00:09.2944072 for: iteration 4
Training: 1.77633832914336
```

It works. MI with Hebbian plasticity works just as well as the regular thing on the char RNN. I want to see how it performs on the full thing.

Before I decide to spent my time optimizing this thing I need to clean it up.

To be honest, I am not sure if there is even a point in expending effort on the full thing since I am just going to be using it for RL.

Since the small char RNN works, the big one should do just as well.

Let me clean it up though.